# Conflicts:
#	.gitignore
#	build.gradle
#	src/main/java/com/example/dplus/alarm/firebase/FCMInitializer.java
#	src/main/java/com/example/dplus/alarm/firebase/FCMService.java
#	src/main/java/com/example/dplus/controller/artwork/ArtWorkBookMarkController.java
#	src/main/java/com/example/dplus/controller/artwork/ArtWorkMainController.java
#	src/main/java/com/example/dplus/controller/notification/NotificationController.java
#	src/main/java/com/example/dplus/controller/post/PostAnswerController.java
#	src/main/java/com/example/dplus/controller/post/PostLikeController.java
#	src/main/java/com/example/dplus/repository/artwork/image/ArtWorkImageRepositoryImpl.java
#	src/main/java/com/example/dplus/repository/artwork/image/ArtworkImageRepositoryCustom.java
#	src/main/java/com/example/dplus/repository/post/like/PostLikesRepositoryImpl.java
#	src/main/java/com/example/dplus/service/account/init/AccountInitialService.java
#	src/main/java/com/example/dplus/service/post/PostMainPageServiceImpl.java
#	src/main/java/com/example/dplus/service/post/bookmark/PostBookMarkService.java

+++ src/main/java/com/example/dplus/controller/post/PostCommentController.java
@@ -34,14 +34,20 @@ public class PostCommentController {
     public ResponseEntity<Success> updatePostComment(@AuthenticationPrincipal UserDetailsImpl user,
                                                      @PathVariable Long post_comment_id,
                                                      @RequestBody PostRequestDto.PostComment data) {
-        return new ResponseEntity<>(new Success("게시글 코멘트 수정 완료",
-                postCommentService.updateComment(user.getUser().getId(), post_comment_id, data)), HttpStatus.OK);
+        if (user != null) {
+            return new ResponseEntity<>(new Success("게시글 코멘트 수정 완료",
+                    postCommentService.updateComment(user.getUser().getId(), post_comment_id, data)), HttpStatus.OK);
+        }
+        throw new BadArgumentsValidException(ErrorCode.NO_AUTHENTICATION_ERROR);
     }
 
     @DeleteMapping("/comment/{post_comment_id}")
     public ResponseEntity<Success> deletePostComment(@AuthenticationPrincipal UserDetailsImpl user,
                                                      @PathVariable Long post_comment_id) {
-        postCommentService.deleteComment(user.getUser().getId(), post_comment_id);
-        return new ResponseEntity<>(new Success("게시글 코멘트 삭제 완료",""), HttpStatus.OK);
+        if (user != null) {
+            postCommentService.deleteComment(user.getUser().getId(), post_comment_id);
+            return new ResponseEntity<>(new Success("게시글 코멘트 삭제 완료",""), HttpStatus.OK);
+        }
+        throw new BadArgumentsValidException(ErrorCode.NO_AUTHENTICATION_ERROR);
     }
 }
